# prof of concept auto avx512 vectorization