DONT'T merge
just example prof of concept for #1591 for output of json_performance

normal code with avx2 code block enabled on znver5

Glaze json write: 0.338783 s, 1886.05 MB/s
Glaze json write: 0.337576 s, 1912.27 MB/s
Glaze json write: 0.231021 s, 3690.51 MB/s

auto vectorized 512 (except movemask_64 )
Glaze json write: 0.318704 s, 2004.88 MB/s
Glaze json write: 0.343197 s, 1880.95 MB/s
Glaze json write: 0.219827 s, 3878.44 MB/s