@stephenberry 2025-02-05
The vector extensions only work for Clang and GCC, right? How would you recommend supporting MSVC?