@arturbac 2025-02-05
> The vector extensions only work for Clang and GCC, right? How would you recommend supporting MSVC?

As i mentioned earlier - [wrap and simulate](https://godbolt.org/#z:OYLghAFBqd5QCxAYwPYBMCmBRdBLAF1QCcAaPECAMzwBtMA7AQwFtMQByARg9KtQYEAysib0QXACx8BBAKoBnTAAUAHpwAMvAFYTStJg1DIApACYAQuYukl9ZATwDKjdAGFUtAK4sGIAMwapK4AMngMmAByPgBGmMQgkv6kAA6oCoRODB7evgFBaRmOAmER0SxxCUm2mPbFDEIETMQEOT5%2BgTV1WY3NBKVRsfGJyQpNLW15nWN9A%2BWVIwCUtqhexMjsHOb%2B4cjeWADUJv5uyGPE4cDH2CYaAILbu/uYRydnBPiC17cPZjsMey8h2ObhihG%2B90eAOerzczWITAAnhDfjsqAcsDQIugIAB9ACyQjcuIAatgAEqLH5QwHAk7hAgXBgAOgQKO2tSU1L%2BTyBLxBeBYLAZTNZ7L%2BrjwVGpkJ56Mx4UwONxuL2hmAKsWR3uB11By8GSM%2BoZADZJKoABy4givAAixsEZutBxVTAIjLwMS8BEwKqgKoAbpgHCRcRkAF6%2B3EQM2LOPHKyQpN3PUHcK0RUHJje1BZhS4rym80WiAu3EsACsXDMeAOAYOWoAtNcHQQzZbnSYAOwWA7ETAENYMOsJo5d20/VPpzPZoh58tVmulwuO4vO%2BtNlsqyvV2vd3v9wfEYcB0fdifJjkZ%2BWYLFKvG4gDikTkxNxVJ1eoNl1b7atNuOe0VzbSRnVdd0Li9H0/QfIMQ2IMM8EjP1Y3jfxExTPVpwiLMc3nYC/1LbdF1rDcDmbfxsF/NcAJ7PsByHA4IAImjFlPdCxwvTDdWwl5Z1zJh8x3JdqI7G0yIoqjiN3McDwY48mOkms2LPcduX8TlMB%2BLwGCyUT/2pHtJzEPBgGYBQY0kLUWPXMQvC0isLAtEwKwndDJyU0jFwTQy3IwlyrEclgFADZAQBAfh1kwXiXInVyfl43C50Egsi0tIiFxkiSWxYsTtW42T8tTVNcv/eiFB8z9ivK5kA0XO0R3cqrisPRj%2BwqpqCuK89qUcgLgtC8LIo2GL4vi%2B5Ev4%2BdhLwZc0rK7LKLLGbCsnPVWoUgNau8zrurU5MptQFJ4jdEgjjME05tXPKmFIfTnRiBsir1fc1t1UrnXayruPa7bqztcthNxQx0FxTAUjwM0ICYP6zFIGJYapXbdQ24cvuRnrZQ0hh8HRXrAosMyvHCsQAHckXzXjSAGsKIpIEaGAzCJYpZ%2B40AYMZMFUFJiGoz7wcwN1cRiREfRLdmxlbMr%2B2Ot0P2417mpR%2BThw0VQNC4DWtc1nXta8WhaAOAAqejZYIb6Xv2h5IT6xyiZJ2hycRSnGcVamQtp4botd5mxq4iWfW53mPptGXBYIYXRcwKGA6l/mzfly2MOq1H7vNnsw6FkWxYgTOCGWU3w8jnO84LvPi8wEtS7u8vs8r3OBbl0g3uq1u29rqOq8b/Oa%2B7iuu4T3uzf7hvB8LrPO9H8OqTU5HOJlO4Q4OJQCC4MGIatLnWBSehcS4CB5cVgrU47n0oY0GINa4a%2Bb6R/yraXle1/BvBN9Ubfd7MA/DOT9aVfw%2BaeIhSLjDAOZ%2BG8IAXyvjfa%2BaF75cSXj6MYwMcbMXms6W6acDgxEPkZZqqcmDnRNNg7654ODLFoJwCsvA/AcC0KQVAnA3DWGsMvVYUVzr%2BB4KQAgmhyHLAANYgDMBaZkXYzAAE4uwWmviaDQXYJESLMBWfQnBJA0L4QwzgvAFAgCCLwuh5DSBwFgEgNALAIb0DIBQCA5jLHDDVEYLgEjNZ8DoD6YguiIAxE0WCZgxBEScG4X45oiIADyMRtDBgMdw8xbBBBhMZoEwxpAsBemAHCA2ujuC8CwCwdU4gUn4H7A4PAQZsn0K5sGb0mxuEMlqJojMMQEQBI8FgTRHoWBBN4EGYgMR0iYFtJgfJRgmagEMcsKgBhgAKBJHgTApMwnHVodw/gggRBiHYFIGQghFAqHUCk3QXB9DqhQCwyw%2BhPS6MgMsI69RsmNjGOgQCphLDWD%2BORMJ/hyL5PWGyfwtpwwMF6RWcitpIi2mwBYOQj5eCoF6RcLA1yD5dGiVkFwONJh%2BGOaERU8xhjHMKJkAQWK9BEvqHMIYCRjl2DRQIXoExPDtD0LS0p9Lxj9DxVSllHLSU0o5ZSioBLlgKHYRsCQFCqEaJSYwjgBxLQmkbGaA4jjgAHGccyLgzINBMVwIQM62wuCLF4AYrQcZSBCLMFqswXYNBWotBafwSiNDyJNCaVRHB1GkFofQ2VOi9E8L4csExiAUCoAsXQeI5BKB2MjQkYAUg4Y0FoB4rxPiUkhICd0%2BG4RQkRKiQ4bNcTGAEESbQZJ9C0leAyWIWg2TuF5IKZsSteASmOHKZoqpyAanZvqZQlJTSWmIjac2k1Fwuk5NIL0/pSghkjOAGMoNfBpmzPmYs5Z2a1nCFEOIbZW69lqE0boOGBgjBnLeRcppyLbkpHuZwR5HwXnnIsB8xsXyfnNGQP8wFwL4igsbOCyF0LYUMIRXgJF8ARW1DpX4CArg%2BXBBxoKhYhL0jEuyEyvIqGihZGQwS1FbKGi8sw9igj9QGWcrKNy/lfQEMzBaHh6lIqxVbMlRwah3rNGyvlRaRVyrgDIGQOqyQzIzC6vwEQXmhrjWBomYI4RKj%2B1ep9XC7RtgA2mqMSG7TZjw32OsTGvTcaQChRSCkXEAZnG4gMEgiOqgzRuJTfENNvjc1ZsnZm8JkTolFvDfE0tSTNFVprVk7NjbRmjtSa2tFHaUldp7ZOvtjTPRDpHR08d2bp0DLnQUy4S6plMBmXMhZSzGCbtkBs3d0h91KEPYcgIJyz2vKsJeq5kGGG3qyNkgA9E889LWLA2cwJLbrYSxPdaoCkCKgkbSjdrN1z9SA7gkgABr1QAxCqFMK4VgYgzcsj6K4OYpI3oXFVGhXUtSGh%2BoCHyW4a5Rdll0HCMUbo898jAqHsodsMR3IpH6OUcGI9o1Kw1jipB/2jjKmtFyvs5IA4NM6waqG5LCAerJOcKNSaoN8nlEeq9V0hRmqNAVgrF2LgFpJBdgrPI/wcNod%2BvU/opdOmQCrAICkb00bbFGasZEVgmw4cI49kjiRzIUcEF4EqfViK9Bbsq1s6rsgD0HPoboZIpMERTZyWxqHXHOBhO9Jzm0qB0RC8R5ZsXEumIeAjVYzh/gZOafNQgQWWAEgoqU7wLpZgNDMgtBIyQkgRFcArGYSQSiTTiM4zKtTujmdyYtQpj1/hpW%2BrU87tjZg0%2BqY4NjxPvSMjOEkEAA%3D%3D)

https://devblogs.microsoft.com/cppblog/avx-512-auto-vectorization-in-msvc/

```cpp
#if defined(__clang__) 
    using uint64x8_t = uint64_t __attribute__((__vector_size__(64)));


    inline auto as_uint64x8( __m512i v ) -> uint64x8_t { return v; }
    inline auto as_m512i( uint64x8_t v ) -> __m512i { return v; }

#elif defined(__GNUC__)
    using uint64x8_t = uint64_t __attribute__((__vector_size__(64)));
    inline auto as_uint64x8( __m512i v ) -> uint64x8_t { return (uint64x8_t)v; }
    inline auto as_m512i( uint64x8_t v ) -> __m512i { return (__m512i)v; }
#else
union uint64x8_t // or memcpy
{
 alignas(64) uint64_t value[8];
 __m512i v512;
};
[[msvc::forceinline]]
inline auto as_uint64x8( __m512i v ) -> uint64x8_t 
    { 
        uint64x8_t res;
        res.v512 = v;
        return res;
         }
[[msvc::forceinline]]
inline auto as_m512i( uint64x8_t v ) -> __m512i { 
    return v.v512;
     }

auto operator &( uint64x8_t a, uint64x8_t b ) 
    {
    uint64x8_t res;
   res.v512 =_mm512_and_epi64(a.v512,b.v512);
   return res;
  }
#endif 
```