@arturbac 2025-02-08
I am not sure it is worth doing so.
When I checked on znver5 generic uint64 version performance it is almost same as optimized versions some cases better some worse.

Glaze json write: 0.336473 s, 1899 MB/s
Glaze json write: 0.390721 s, 1650.43 MB/s
Glaze json write: 0.200108 s, 4260.62 MB/s